Refactor pages to render roadmap cards as a separate component
